+++ dex/src/main/scala/com/wavesplatform/dex/actors/orderbook/OrderBookActor.scala
@@ -9,8 +9,9 @@ import cats.syntax.apply._
 import cats.syntax.option._
 import com.wavesplatform.dex.actors.MatcherActor.{ForceStartOrderBook, OrderBookCreated, SaveSnapshot}
 import com.wavesplatform.dex.actors.address.AddressActor
+import com.wavesplatform.dex.actors.events.OrderEventsCoordinatorActor
 import com.wavesplatform.dex.actors.orderbook.OrderBookActor._
-import com.wavesplatform.dex.actors.{MatcherActor, OrderEventsCoordinatorActor, WorkingStash, orderbook}
+import com.wavesplatform.dex.actors.{MatcherActor, WorkingStash, orderbook}
 import com.wavesplatform.dex.api.ws.actors.WsInternalBroadcastActor
 import com.wavesplatform.dex.api.ws.protocol.WsOrdersUpdate
 import com.wavesplatform.dex.domain.asset.AssetPair
