+++ waves-integration/src/main/scala/com/wavesplatform/dex/collections/ListOps.scala
@@ -1,4 +1,4 @@
-package com.wavesplatform.dex.collection
+package com.wavesplatform.dex.collections
 
 import scala.annotation.tailrec
